Update umbraco-marketplace.json
Hi Jason - we're about to introduce a new category on the Marketplace for "Artificial Intelligence" and from reading the details of your package, I think it would make sense to list it in this category - as a secondary one alongside the existing main category.

If you are in agreement, I've raised this PR to do that.  If accepted and merged, your package when synchronised will be listed in the new category once it's published.

Please just note that as you were already using the "alternative category" field to list this in "Developer Tools", if you accept this PR it will be removed from that one (we only support listing in a maximum of two categories).  So it's totally up to you what you prefer.

Thanks for your support of the Umbraco community with your package.